Add plot.py script for generating and saving plots
This commit adds a new script called plot.py to the postprocess
directory. The script is responsible for generating and saving plots
based on the data provided. It uses matplotlib library for plotting and
pandas for data manipulation. The generated plots are saved in the
specified output directory. The script also accepts a command line
argument to enable log scale for the y-axis if needed. This addition
will enhance the postprocessing capabilities of the project.